Fix stack overflow when running `provision` with `--output json` (#1527)
The implementation of some MarshalJSON provided by some of our UX
Objects ended up cuasing infinate recurison becuase the implementation
passed the object being serialized as the `Data` property of the
`EventEnvelope`. This meant that we'd end up trying to serialize the
same object over and over again (in a nested envelope structure) until
we ran out of stack space.

To fix this, I adopted a similar pattern that we were using in the other
UX Objects, i.e. send a "console event" message that contains the output 
that would have been written to the console.

This also fixes a case where `azd show` would panic if there was a
java based services.

I also modified one of our end to end tests to do an infra create using
`--output json` so we have more end to end coverage in CI with JSON
output.

Fixes #1437