Fix environment flag not respected (#1534)
This changes the resolution of environment flag to bypass IoC registration and use the hardcoded `environmentName` flag. A test has been added to verify that `-e` is respected in the CLI. This test failed prior to these changes.

Explanation: With the switch to `golobby`, there isn't an ability to specify bindings between concrete and interfaces except through the resolver function's return type. A concrete flag instance is being registered, and thus in the existing code, only the concrete type is resolved, and `flags` is never resolved. Switching to named registration is also not an option, since: we need a single unique name, and actions end up sharing the same container. Thus, the second time a `flagResolver` is registered, it will fail.