Remove service calls for location and subscription defaults in UI prompt (#1508)
Currently, the concept of a `default.location` and `default.subscription` is tied to the UI selection. For the UI selection, it's not important to validate the default configuration's validity, and it is preferred that the UI simply chooses no default, or a hardcoded default, if the default configuration isn't valid.

This change removes the extra validation we're performing to validate account defaults. The API and implementation for validated defaults is still available in case we want better guarantees for other use-cases.