Fix tests that verify command and actions initializing correctly (#1542)
Fix the existing test to correctly verify that commands and actions can be initialized (without running).

In the previous code, we construct a root command to discover all possible runnable commands. We then call `Execute` on a child command. This actually resulted in the root command being executed without args, i.e. simply calling `azd`. The relevant code in `command.go` of cobra:

```golang
	// Regardless of what command execute is called on, run on Root only
	if c.HasParent() {
		return c.Root().ExecuteC()
	}
```

To fix this, a new command instance without `Parent` is created with the appropriate args, then executed.