Add tags to ARM deployments
This change updates the Bicep provisioning provider such that
deployments we create are tagged with both the name of the azd project
they came from as well as the environment that was used for the
deployment.

Having this metadata avaliable will be useful in the future as we
start changing the name of the deployment object we create from just
the environment name to a combination of the environment name the
project name and a timestamp (so a seperate deployment object is
created per `azd provision`). When we do that, this will provide a way
to find the deployment object (or objects) that correspond to a
specific project/environment pair, without having to do name matching.

It can also be useful for querying purposes.