Update terraform provider versions (#1717)
As described in issue #1716 the Terraform provider version in the templates is not supporting Node.js version 18-lts. 
This PR updates the `azurerm` provider to the latest available minor version and this way fixes issue #1716 .

In addition, this PR updates the version of the `azurecaf` provider. This is more a "cosmetic" update as due to the version restriction in the Terraform files this version was fetched anyways.