Remove environment implicit effects on builds
Currently, when `azd` packages an application, for some specific languages, we thread the environment through. This is not a consistent behavior, and not the behavior we want, as it leads to non-reproducability across builds.

We may reintroduce this functionality as a separate gesture in the near future, but we should not auto the current behavior as it will lead to unexpected behaviors in production apps.