Call Environment.GetSubscriptionId in fewer places (#1748)
* Call `env.GetSubscriptionId()` in fewer places

In a bunch of places in our code we are interact with a resource where
we have the resource id. This contains the subscription id, so we can
use this value directly instead of reading it from the Environment.

* Reduce AzureResourceManager dependency on Environment

Be more explicit about what information we need (the subscription id
and environment name) and encode that in the API design for
AzureResourceManager. It feels more like an SDK client now.