Move `SubscriptionsService` into `account` Package
As part of some future work, I need to be able to take a dependency on
the `auth` and `account` packages from `azcli`. That can't be done
today because some code in the `account` package depended on the
`azcli` package and so what I wanted to do would lead to an import
cycle.

Looking at the code, it seemed like the best thing to do as move the
`SubscriptionsService` out of the `azcli` package and into
`account`. This change does the refactoring of doing that.

I tried to this as non destructively as possible to the existing
implementation, but did need to move away from some of the model types
in the `azcli` package (which I think long term we should just remove)
in favor of the types exposed by the armsubscriptions SDK.