Don't register a `TokenCredential` in our container (#1763)
The TokenCredential we produce is tied to the subscription that is
registered in the environment (so we ensure we use the tenant id that
matches the subscription). This works okay-ish today because we know
once `init` has run that the environment will contain a Subscription
ID that we can use.

I plan to break that invariant in the future, delaying asking about a
subscription until `provision`. In that world, it is better if
components like the `AzCli` take a component that allows for
requesting a token to use for a given subscription.