Make `Test_MavenProject` tests work when `mvn` is missing
These unit tests were failing on my local machine because I do not
have Maven installed globally. It seems like we could get these tests
to work if we just orchestrated the test folder such that the Maven
wrapper was present in the `api/src/` folder (as we have in our
templates).

This change updates the tests such that we now lay down an `mvnw`
"script" (it's actually a zero byte file, but good enough to convice
`os.LookPath` that the wrapper is there and ready to be used) and then
updates the test to expect that we invoke `mvnw` instead of `mvn`.

The diff here looks a little noiser than it is, because I made each
scenario (Restore, Build, and Package) a subtest of
`Test_MavenProject` so I could share the setup logic that creates the
`mvnw` wrapper. The only change to the logic of the tests is to change
an assertion to validate that we are trying to run `mvnw` instead of
`mvn`.