Ask fewer questions during `init` (#1752)
This change reworks environment initialization to no longer require a subscription or location.  Instead, these properties are now state that are managed by the individual providers and stored in the environment (long term, we envision other provisioning providers which don't need a subscription id or location, which is why each provider manages this state instead of the `provisoning.Manager` type).

Since `init` and `new` no longer configure either the subscription or location, they no longer require you to be logged into Azure.

In addition, `AZURE_PRINCIPAL_ID` is no longer added to the `.env` file.  Instead, the value is present when evaluating the parameters template file (this means that if you log in as a different principal from the one you created the environment with, when you run `azd provision`, you'll get the id of the current principal, not the old principal, which makes more sense) but not persisted to disk.

Components which need to call either `GetSubscriptionId` or `GetLocation` must now guard against the case where these values are not set, because `provision` (or `env refresh`) have not been run yet. In these cases, we tell the user the command can not run because infrastructure has not been provisioned and guide them to run `azd provision`.