Move `postpackage` to `predeploy` hooks (#1856)
Move `postpackage` to `predeploy` hooks for `staticwebapp` ToDo templates

This allows the `env-config.js` in the ToDo templates (environment configuration) to be correctly generated when observing the environment state from `azd`. This is the current intended integration for `staticwebapp` to observe environment state that also works when invoked through `azd up`.