Support other java archive files (#1931)
Relax packaging validation to also allow for `*.war` and `*.ear` files in addition to `*.jar` files. Provide ability to specify a specific archive file if multiple archives are present.

Functionality-wise, these archive file formats are transparent to `azd`. For `zipdeploy` deployments, they are strictly zip files that simply needs to be uploaded to the target endpoint. This has been tested end-to-end working by reliable-web-app-pattern-java team.

Fixes #1943