Add `requiredVersions` to `azure.yaml`
This allows a project to specify a range of supported versions of
`azd`, by introducing a new `requiredVersions` object to
`azure.yaml`. Today, the only supported key is `azd`, which allows the
project author to specify a set of supported versions.

When parsing a project file, if this member is set, we ensure that the
current version is within the allowed range.

Dev versions of the CLI ignore this check.

Fixes #1504