Add `requiredVersions` to `azure.yaml` (#1704)
* Fail fast when embeded version is malformed

This change reworks our internal logic around how we store and report
the version of `azd`. The goal is to centralize the validation logic
inside the `internal/version.go` so the rest of the system can assume
there is valid version information and not worry about error handling
paths.

* Add `requiredVersions` to `azure.yaml`

This allows a project to specify a range of supported versions of
`azd`, by introducing a new `requiredVersions` object to
`azure.yaml`. Today, the only supported key is `azd`, which allows the
project author to specify a set of supported versions.

When parsing a project file, if this member is set, we ensure that the
current version is within the allowed range.

Dev versions of the CLI ignore this check.

Fixes #1504