Make error message better when no subscriptions are found (#2004)
When no subscriptions are found, the user is currently presented with an unfriendly error: `please provide options to select from`

This fix improves the error message and provides an error hint, and also improves other similar occurrences of `console.Select` without length checking on options. This change also updates the error formatting logic to only colorize the first line, in line with our desired UX patterns.