Store encrypted MSAL cache with algorithm type (#2031)
When storing the MSAL cache on Windows we use CryptProtectData to
ensure that the data is encrypted at rest. Previously, we stored just
the raw data as returned by CryptProtectData. This would make it
difficult in the future for us to change our encryption strategy
because we would be unable to determine from the opquae encrypted data
how it was encrypted (and hence what we needed to do to decrypt it).

This change updates our strategy to encrypt the data as before, but
then store it as part of a JSON object which contains a string
denoting how the data was encrypted.

To support transparent upgrades, care is taken to fall back to the
older strategy in the case where we can't succesfully unmarshall the
stored data as a JSON object (this allows new versions of `azd` to
read the cache written by older versions). Future writes to the cache
will store the data in the new format.

This introduces a slight breaking change where if you upgrade to a
version of `azd` that has this change and then downgrade to an earlier
version, the earlier version will not be able to use the cache. In
practice, this will cause the CLI to behave as if it was logged out
(since all our auth logic will fail) and the user will need to `azd
login` again with the older version, which will recreate the cache in
the older format that it can read.

Fixes #1545