Move "current-user" settings out of `config.json`
Previously, the identity of currently logged in user (either the home
account id, when logged in as a user, or the client and tenant id when
logged in as a service principal) as stored in user's `config.json`.

A downside of storing this information in `config.json` is that when a
user did `azd config reset`, it would "log out" the current user,
which is very surprising behavior. In addition, anyone looking at
`config.json` would see properties that were much more like state of
the auth subsystem instead of actual config values that a user would
want to change.

So, this change moves this information out of `config.json` and
instead stores it in `auth.json` (a new file) which is also stored in
`~/.azd`.

On first use, we'll migrate the relevent properties from `config.json`
into `auth.json`.

Fixes #1721