Use a unique deployment name when provisioning
Previously, when creating a deployment, `azd` would use the
environment name as the name of the deployment (a deployment is
represented in azure as a resource, and has a name).

When running `azd provision` twice, the second invocation would re-use
the previous deployment name, which would overwrite the history of the
old deployment.

This was not ideal, since we would lose all the history from the
previous deployment and could also lead to race conditions in our UI
where we would end up displaying information from the previous
deployment while the new deployment was waiting to start up (since the
deployment name was shared we would fetch the operations from the old
deployment, not the new one).

Also, in the case of subscription based deployments, the deployment
object also has a location tied to it, and reusing the name meant that
if you changed `AZURE_LOCATION` for an environment, and tried to
`provision` again, you'd get an error from ARM saying that a
deployment with that name already existed in another region, which is
very hard to understand.

This change moves to using a unique name per deployment, by taking the
current unix time and appending it to the environment name. We also
now tag the deployment we create with `azd-env-name`, which is used to
find the most recent deployment for a given environment, for example,
when `azd env refresh` is running.

As part of this work, the existing `Deployment` interface has been broken
into two parts, a `Scope` which lets you do operations across a scope
and `Deployment` which represent a named deployment tied to a specifc
scope.

Ideally, we would move `infra.Scope` and `infra.Deployment` out of the
`infra` package, since they really are Azure specific concepts, but
that's a task for another PR.

Fixes: #293