Remove `Scope` parameter from `infra.Provider` and `infra.Manager` (#2082)
The `Scope` concept is something which is very ARM specific. The fact
that it was a prameter on `Manager` and `Provider` is an artifact of
how the interfaces were birthed (by refactoring the at the time only
provider we had, which talked to to ARM) and the the fact that we were
doing ARM deployments for ACA.

We are now at a point where we can remove it from the interface, so let's
go ahead and do that.