Ensure generated deployment name is not too long
Since we started to append a timestamp to the deployment name we ran
into an issue where it was possible to generate a deployment name that
was longer than the ARM allowed limit.

With this change, we now truncate the generated name to length (taking
the last 64 chracters of the string, on the assumption that the
"interesting" parts of the environment name are near the end of when
you have a very long name).

The deployment object is still tagged with the full environment name
so that `azd env refresh` and others who want to query by full
environment name can find it.