Allow Resource Group Scoped Deployments (#2083)
This change adds basic support for using `azd` with bicep templates
that target resource group scope instead of subscription scope.

This support is guarded by the `alpha.resourceGroupDeployments` flag
while we iterate on the UX and general design, so to enable it, one
must run `azd config set alpha.resourceGroupDeployments on` to enable
the feature.

Like all experimental features, we issue a warning when they are used
to remind the user that they are using an alpha feature that may
change over time.

I also added `AZD_DEBUG_NO_ALPHA_WARNINGS` which may be set to 1 or
another value as understood by `strconv.ParseBool` which will supress
these warnings.

With these changes, we now examine the compiled ARM template to
determine the scope it targets. In the case of a resouce group targed
deployment, we expect that `AZURE_RESOURCE_GROUP` is set in the
environment and we use this to determine what group to deploy the
template into. We'll build a real UX that allows the user to either
select an existing resource group or create a new one shortly.

The behavior of `azd down` with this change is to attempt to delete
the resource group that was deployed to (this is an artifact of the
existing implementation and when exposed to resource group scoped
deployments, but it seems like a pretty reasonable way to start)