Provide a sugesstion for an environment name during `init` (#2134)
This change provides a suggestion for the initial environment name
that the user needs to create during `azd init`.  The suggestion is
based on the name of the current directory, suffixed with `-dev`.

In addition, some help text is added to help give users some idea as
to what this value is used for.

I had hoped that we would be able to default this value to something
like `main` and simply not ask the question to the user, but that is
more difficult than expected, given how `azd` is used today. Most
existing `azd` templates assume that `AZURE_ENV_NAME` is going to be
used in a way that forces it to be unique across the user's
subscription (since it is used to both create the name of the resource
group we provision resources into and used to generate the resource
suffix that we append to most resources to get unique names)

In addition, even when starting from scratch, our system wants you to
tag your resource groups and resources with an `azd-env-name` tag so
that other parts of `azd` can discover these resources. So even if we
hid the concept during `init`, you'd have to learn about an azd
environment as soon as you start interacting with your IaC.

Given that, the best we can do to ease the pain for folks getting
started is to provide a reasonable default, and using the name of the
current directory (suffixed with something like `-dev` since there
can be multiple environments per project) feels reasonable.

Fixes #335