Move tracing logic to `tracing` package (#2203)
This change moves logic that used to be part of `telemetry` into `tracing`. This allows separation of application tracing logic from telemetry initialization and upload logic. It also allows nicer gestures for tracing vs. logging events.

As an example, `telemetry.GetTracer().Start` is now available as `tracing.Start()` which is much more readable.