Support environment name as explicit argument for `env refresh`
With this change, `azd env refresh some-environment-name` is now
supported and behaves the same as if the user had typed `azd env
refresh -e some-environment-name`.

When an explicit environment name is not provided via the `-e` or
`--environment` flag and there are no arguments, `azd env refresh`
refreshes the default environment, as before.

Fixes #1746