WIP: Use `os.Getenv` instead of accessing `Environment.Values`
This allows us to inherit values from the OS Environment which is nice
behavior and also supports CI/CD cases where we set values which are
normally stored in the .env file as environment variables (because the
.env file is not persisted across CI runs).

For folks that care about getting at just the values in the `.env`
file stored with the environment, `Environment.Dotenv()` now returns a
copy of it. The most common use of this is tests which want to
validate that the environment file was updated in someway.

Fixes #2208