Return failed error when logging in (#2243)
Previously, when `login` fails, there are cases where "Not logged in, run 'azd auth login' to login" would be printed out, but the returned exit code is zero. This is a problem in CI situations when auth isn't setup correctly, and yet `login` is deemed successful, the CI pipeline proceeds as normal, resulting in user confusion.

This change updates `login` to always exit with non-zero exit code if `login` fails. The code has been refactored so that `--check-status` plays nicely with this implementation.

Fixes #2221