Allow ARM native installs when explicitly passing `-a arm64`
This change updates install-azd.sh to support installing the arm64
version of `azd` when the user explicitly passes `-a arm64`. When run
without the `-a` flag, `install-azd.sh` continues to install the amd64
version of macOS and fails on Linux, as before.

This allows someone to use the following line in their Dockerfile to
install `azd` when creating a dev container:

```
RUN curl -fsSL https://aka.ms/install-azd.sh | bash -s -- -a $(dpkg --print-architecture)
```

To install either the amd64 or arm64 version of based on the
architecture of the underlying system.