Improve resource group onboarding
Previously, if a bicep template targeted resource group scope, we
required that you set `AZURE_RESOURCE_GROUP` in the environment to
determine what resource group you wanted to deploy to.

With this change, we now instead prompt you when this is unset to
either select an existing resource group based on the ones in your
subscription, or create a new resource group.

When creating a new resource group, we provide a default value of
`rg-<environment-name>` (matching our templates) and when we create
the resource group, we tag it with `azd-env-name` as is our
convention.

Existing resource groups are used as is, with no changes.

The name of the resource group that you have targeted is stored in the
`.env` file as `AZURE_RESOURCE_GROUP`, similar to how
`AZURE_SUBSCRIPTION_ID` and `AZURE_LOCATION` are.

When `azd down` is run, `AZURE_RESOURCE_GROUP` is removed from the
`.env` file, since `azd down` deletes the resource group as part of
its operation.