Remove `httputil.HandleRequestError` (#2316)
Remove `httputil.HandleRequestError` and its usage as it may be more correct to return the underlying `err` when `err != nil && response != nil`.

Reasoning is as follows: suppose that the transport, or an executor of the pipeline, had returned an `err` for some reason, with `resp` intact. The executing code has made a decision that the request should be considered an error. The server response, in this case, is not the cause of the error. Instead, something else has occurred.

For example, it is documented in [http.Client.Do](https://pkg.go.dev/net/http#Client.Do):

> On error, any Response can be ignored. A non-nil Response with a non-nil error only occurs when CheckRedirect fails, and even then the returned Response.Body is already closed.

In this case, it is more correct to attribute the error to the maximum redirects reached, and not the HTTP response with StatusCode 307/308 from the server.