Store numeric values with leading zeros in `.env` correctly
The package we are using for storing values in a `.env` file has some
surprising behavior where numeric like values are stored as unquoted
strings with any leading zero dropped (since the value is round
tripped via `strconv.Atoi` and then `fmt.Sprintf` with a `%d` format
directive.

This leads to issues where if you have a bicep file that does something like:

```bicep
output NUMBER_WITH_LEADING_ZERO string = '01'
```

The value stored in the `.env` file will be just `1` instead of `01`.

This change addresses this issue by detecting the case and ensuring
that we store the raw value in the .env file (after wrapping it with
quotes).

Fixes #2307