Allow `show` to work when no environment exists
Today, `azd show` is used by VS Code to get information about the
current project, both the services that make it up (and information
about where the code for these services is) as well as the target
resources (if we can determine them)

In the case where an environment didn't exist (which is hard to have
happen during a normal course of action, but possible if you manually
go mucking around in `.azure`) `azd show` would prompt the user if
they wanted to go create an environment.

This caused issues with the VS integration, since the editor would not
expect `azd` to prompt in this case.

`azd show` already handles the case where the infrastructure for a
project has not been deployed. This change extends this behavior so
that if for some reason we can not fetch the current environment, we
just return information about the services that we can discover

Fixes #2296
Fixes #851