Skip update check in test that ensures `--help` is well formed (#2421)
Skip update check in `Test_CLI_NoDebugSpewWhenHelpPassedWithoutDebug` that ensures `--help` is well formed.

Background: the test is asserting that `stderr` isn't polluted by unexpected debug spew when a simple `--help` is passed. The version update check is one of the things that could write to `stderr` but is outside the purpose of the test.