Refactor Docker `Build` to take an `io.Writer`
This changes the design such that `Build` now takes an `io.Writer`
that is used to write the output of `docker build` to. This means that
`Docker` no longer needs to hold a reference to console or be
concerned with setting up the previewer.

That responsibility is now in the docker framework service, which is
also wrong long term (since it means that the framework service needs
a reference to the console, which feels incorrect) but that battle can
be fought another day.