Set CWD when launching `azd` to upload telemtry (#2467)
When `azd` is about to exit, if there is telemetry that needs to be
uploaded, we spawn a seperate copy of `azd` to go do the uploading and
then exit the current process.  This allows the upload of the
telemetry to happen in the background instead of pausing the
termination of `azd` until it is complete.

Previously, we reused the current working directory that the root
`azd` process was launched in (which, in the common case is whatever
directory the user ran `azd` from).

On Windows, if a process is running with a given CWD, that directory
can not be deleted until the process exits. This means that until the
upload process finishes and the uploader exits, the CWD of the initial
`azd` is in use and can not be deleted.

In CI our end to ends tests run in an ephemeral directory, created by
`(*testing.T).TempDir()`, and `azd` is run from within this
folder. When the test completes, the testing library tries to clean up
this temporary directory, and this can fail (depending on timing)
because the directory is in use, since the uploader is running.

To address this issue, when we launch `azd` to upload telemetry, use
the directory that `azd` is contained in, instead of what our current
working directory is.