Move `EnsureEnv` from `prompt.Prompter` to `provisioning.Provider` (#2465)
Different provisioning providers may require different values to be
set in the environment to be able to do a provision operation.  For
example, when using the bicep provider and doing a resource group
scoped deployment, in addition to a subscription and a location, we
need `AZURE_RESOURCE_GROUP` defined to name the target resource group to
deploy to.

This change prepares us to have providers require different values to
be set in the environment by moving `EnsureEnv` from `prompt.Prompter`
and to the providers themselves.

There should be no functional impact of this change, it is simply code
motion to help with some future changes.