WIP: Support fetching credentials from an external process
This change sketches out a possible way for `azd` to delegate
authentication to an external process and teaches the VS Code
extension to use this strategy to provide authentication for `azd`
based on the credential associated with the Azure Account extension.

This works by introducing a new "Remote Credential" type in `azd`, is
used when `AZD_AUTH_ENDPOINT` and `AZD_AUTH_KEY` are defined in the
system environment.  When these values are set and `azd` needs to
request a token, it makes a request to
`${AZD_AUTH_ENDPOINT/token?api-version=2023-07-12-preview` using
`${AZD_AUTH_KEY}` as an bearer token to secure the request.

The body of the request body contains a serialized version of the
`TokenRequestOptions` struct:

```
{
   scopes: []string
   tenantId: string?
}
```

The server is expected to respond with 200 OK and a body that
represents the result of trying to fetch a token (either success or
failure).  One of the two following shapes is expected for the
response:

```
{
  status: "success"
  token: string
  expiresOn: string // (RFC3339 format)
}
```

or

```
{
  status: "error"
  code: "GetTokenError" | "NotLoggedInError"
  message: string // Suitable for display to a human.
}
```

This behavior is currently controlled by a setting
`azure-dev.auth.useIntegratedAuth` which defaults to false. You need
both the updated extension and a version of `azd` which supports this.

Contributes To: #1741