Fix `azd down` after a failed `azd provision`
For subscription level deployments, `azd down` needs to discover the
set of resource groups from the previous deployment, so it can go
delete each one. We were using the `outputResources` property of an
ARM Deployment to compute this, which is not populated for failed
deployments.

In the failed deployments case, we revert back to trying to infer this
information (as before d8863e44c25bad9e52a5c6a1c0b042a52b81edde) from
the `dependsOn` portion of the deployment. This can miss some cases
(e.g. empty resouce groups, which is why we moved away from this in
failure cases (where often you really do want to tear down everything
and start over)

Fixes #2508