Fix asker prompting indefinitely due to invalid choice selection on non-tty mode (#2565)
This change fixes asker to error out if an invalid choice is provided while reading in non-tty mode.

Context: while reading a `survey.Select` prompt choice from a non-tty stdin (such as a file, or `strings.Reader` in tests), an invalid choice causes the application to stall indefinitely. This is due to the use of a for-loop, where `asker` was re-prompting for invalid input instead of erroring out.

The `readStringNoBuffer` implementation has also been modified slightly to handle a non-zero byte read with io.EOF encountered, which is possible as described by io.Reader.