azd package support for user specified output paths (#2721)
Adds ability for a user to specify a --output-path param to the azd package command to control the output location of file based packages.

For top level commands it would be expected that user specify a absolute or relative folder path. If the path does not exist azd will create it automatically.