Build cmd.exe command line directly
When using the shell to run a command on Windows, we were not properly
quoting arguments that we passed to cmd.exe. This could cause errors in
cases where paths to command or arguments contained parenthesis.

To address this, we no longer rely on Go's quoting of arguments and
instead set the CmdLine field of SysProcAttr on Windows, with a command
line quoted as `cmd.exe` expects.

As part of this change, I also fixed an issue where we were not flowing
the context object to the subcommand when we were using the shell to run
the command instead of invoking it directly.

A small end to end test is included to ensure we can invoke programs
when they live in directories that have `()`'s in their names, as we we
saw in #2727 where the downloaded version of bicep is cached into the
user's folder, and the user name contained `()`.

Fixes #2727