WIP: Support Multiple Resource Groups
Allow using the `SERVICE_<SERVICE_NAME>` pattern to specify the
resource group for a specific service. This can be useful when you
allocate multiple resource groups and want to deploy different
services to different resouce groups.

Longer term, we should probably update our logic such that when trying
to find a target resource, if there are multiple groups we list each
and try to find the resource with the correct tag, but this is a quick
hack to unblock some folks who are exploring using `azd` in these more
complex scenarios.

Rasing a PR mostly so I can get privates to share with them