Fix spurious test output (#2836)
Fix spurious test output in CI.

In CI, we see these specific tests reported as FAILED in the log, but not actually failing. The reason is that `gotestsum` is unable to find the actual "PASS" event emitted, but `go test` does succeed correctly. We have trailing output that isn't newline terminated (while waiting for stdin), which results in the behavior.

Example:
```
=== FAIL: cli/azd/internal/repository TestInitializer_infraSpecFromDetect/api_and_web_with_db (unknown)
Input the name of the app database (PostgreSQL) --- PASS: TestInitializer_infraSpecFromDetect (0.00s)
```