feat: clickable url for azd template show (#2867)
The command azd template list gt the feature of clickable repository paths in release 1.4.2. This PR brings this feature also to the command azd template show xyz transfering the logic implemented in #2845 to the show command.

To avoid code duplication the formatting logic is trenafered to the path.go file which looks like a good fit for this kind of logic.

The sequence of the output of the template show command was not changed. If this should be aligned with the template list command, just comment and I add a commit