Add `project.DotNetContainerAppTarget`
This adds support for a new (presently internal) service target, which
is a containerized dotnet application, running on ACA. There are two
major differences between this target type and the existing container
app type:

1. Instead of using `docker build` and `docker push` to build and push
the container, we use `dotnet package` and the new container building
support that is in .NET 7 (and being improved in .NET 8) to build and
push the ccontainer. This experience is like Oryx in that it does not
require a Dockerfile.

2. We use a deployment strategy for the container that looks more like
AKS. We expect a `containerApp.tmpl.yaml` file in a `manifests` folder
in the root of the service. We use go text templating (as we do in
AKS) and then deploy this yaml in the same manner as `az containerapp
[create|update] --yaml`. The format of the file is the same between us
an `az` (it is just the YAML rendering of what would be sent to the
management plane to create or update a container app)

We do not allow this type service target type to appear as a host in
`azure.yaml` today. Instead, it will be returned by a custom importer
in a later change.

We will explore extending this support to other languages and at that
point consider exposing this via `azure.yaml`.