Support .NET App Host Projects (#2956)
.NET App Host projects can produce a manifest which `azd` can use
instead of our existing huristics to determine what resources an
application needs.

This change adds support for these projects to our "initialized from
an existing codebase flow".

In addition, this feature uses the new support in `azd` for not
persisting IaC to disk by default. A new `azd infra synth` command
(presently in alpha) is added for folks that want to persist the
infrastructure to disk, either to look at it or modify it if there are
issues.