Handle if project is undetected (#2963)
Add a nil-check to handle when a python project is undetected. In this case, the python project will still be attempted to be built by the Oryx builder. This would likely result in a different, but friendlier error message and avoids the panic.

With this change, better error messaging is also provided when building a containerized project fails due to no Dockerfile and failed detection.

Fixes #2961