WIP: Improve testing for Aspire apps
This change improves test coverage for Aspire applications. There are
a set of unit tests for the `apphost` package, focused on generation
(including some baselines which should allow us to do a beter job of
understanding how changes to our genertion logic impact the generated
code) and an end to end test that deploys the aspire-starter template
from `dotnet new`.