Do not require `docker login` for Aspire apps
The `dotnet` toolchain is able to build and push container images for
Asprire apps via, `dotnet publish`. However, we still needed `docker`
around so we could run `docker login` to log into the private ACR
registry we intended to push the image to.

To remove the dependency on `docker login` we can use the
`SDK_CONTAINER_REGISTRY_UNAME` and `SDK_CONTAINER_REGISTRY_PWORD`
environment variables as described in:

https://github.com/dotnet/sdk-container-builds/blob/main/docs/RegistryAuthentication.md#authenticating-to-container-registries

The caveats here do not apply to us. If they did, we could consider
either editing the `~/.docker/config.json` directly, to add our auth
information or setting `DOCKER_CONFIG` when running `dotnet publish`
to a version that we write (just in time, in a temp dir) and clean up
after.

Note that the credentials fetched via the AAD flow are shortlived
(~3hr of validity) and so caching them in `~/.docker/config.json` long
term is less than ideal (our documentation for commands like `az acr
login` mention that you should do this before each operation).